refactor: 기본시세 실전 domain Controller import 제거, Service 시세분석 메소드 분리
